>	Added a bunch of [[maybe_unused]] attributes to silence compiler warnings when 	I tried using highest warning level
>	Fixed one serious bug I had found with String::InsertAt because of these warnings

+++ Library/Sources/Stroika/Foundation/Characters/String.h
@@ -1077,7 +1077,7 @@ namespace Stroika {
                  *
                  *  This is only used for 'STL-compatabiliity' apis, like substr (), find, rfind (), etc.
                  */
-                static constexpr size_t npos = -1;
+                static constexpr size_t npos = static_cast<size_t> (-1);
 
             public:
                 /**
